[river-api] Split batch, preview and scripts Views to different files
