+++ django/control/api/router.py
@@ -1,10 +1,10 @@
 from rest_framework import routers
 
-from control.api import views
+from control.api import batch, preview, scripts
 
 # Register your API views here
 
 router = routers.SimpleRouter()
-router.register(r"batch", views.BatchEndpoint, basename="batch")
-router.register(r"preview", views.PreviewEndpoint, basename="preview")
-router.register(r"scripts", views.ScriptsEndpoint, basename="scripts")
+router.register(r"batch", batch.BatchEndpoint, basename="batch")
+router.register(r"preview", preview.PreviewEndpoint, basename="preview")
+router.register(r"scripts", scripts.ScriptsEndpoint, basename="scripts")
